# api/openapi.yaml
openapi: 3.0.3
info:
  title: MoonLightCloud API
  description: >-
    Device lifecycle + versioned sync ingestion with the first Sale-sync
    vertical slice: durable inbox, sale.finalized.v1 validated before ACK,
    async projection with durable processing state. Business APIs are
    versioned under /api/v1. Health and version stay outside the version
    boundary. ACK means durable PostgreSQL commit, never queueing alone.
  version: 2D
paths:
...
  /api/v1/device-control/poll:
    post:
      summary: Retail control poll (lease at most one due command, refresh presence)
      operationId: deviceControlPoll
      security:
        - deviceAuth: []
      responses:
        '200':
          description: Server time plus at most one leased command.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceControlPollResponse'
        '401':
          description: Missing, unknown, revoked, or invalid credential (presence untouched).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/device-control/commands/{id}/accepted:
    post:
      summary: Report durable Retail receipt (idempotent)
      operationId: deviceControlAccepted
      security:
        - deviceAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current command state after accepted receipt.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceControlCommandResponse'
        '401':
          description: Missing, unknown, revoked, or invalid credential.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown command or cross-device lookup.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/device-control/commands/{id}/status:
    post:
      summary: Report running/completed/failed (first terminal wins)
      operationId: deviceControlStatus
      security:
        - deviceAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceControlStatusReport'
      responses:
        '200':
          description: Current command state after the report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceControlCommandResponse'
        '400':
          description: Malformed body, unknown status, or invalid result code (DEVICE_COMMAND_INVALID_TRANSITION).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing, unknown, revoked, or invalid credential.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown command or cross-device lookup.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Contradictory terminal outcome (DEVICE_COMMAND_CONFLICT).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/operations/incidents:
    get:
      summary: Bounded incident list with state/severity/rule filters
      operationId: operationsIncidents
      security:
        - dashboardSession: []
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, acknowledged, resolved]
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [warning, urgent]
        - name: rule
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: One bounded page plus next_cursor (empty when final).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsIncidentList'
        '400':
          description: Invalid filter, limit, or cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/operations/incidents/{id}:
    get:
      summary: Incident detail with masked deliveries and recovery state
      operationId: operationsIncidentDetail
      security:
        - dashboardSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Incident with deliveries (recipients masked) and recoveries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsIncidentDetail'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown incident.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/operations/incidents/{id}/acknowledge:
    post:
      summary: Acknowledge an incident (idempotent, never resolves)
      operationId: operationsAcknowledge
      security:
        - dashboardSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current incident state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsIncidentResponse'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown incident.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/operations/incidents/{id}/resolve:
    post:
      summary: Resolve an event incident, or a cleared stateful condition
      operationId: operationsResolve
      security:
        - dashboardSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current incident state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationsIncidentResponse'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown incident.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Stateful condition still active (CONDITION_STILL_ACTIVE).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/operations/summary:
    get:
      summary: Aggregate counters and batched device incident rollup
      operationId: operationsSummary
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Low-cardinality counters plus per-device aggregates.
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/operations/devices/summary:
    get:
      summary: Per-device active incident counts in one batched read
      operationId: operationsDeviceSummary
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Device rollup array (never per-device queries).
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/devices:
    get:
      summary: Operator device connectivity + latest sync command
      operationId: dashboardDevices
      security:
        - dashboardSession: []
      responses:
        '200':
          description: Devices with NEVER_SEEN/ONLINE/OFFLINE connectivity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardDeviceList'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/catalog-admin/commands:
    get:
      summary: Store-scoped catalog command history (newest first)
      operationId: catalogAdminListCommands
      security:
        - dashboardSession: []
      parameters:
        - $ref: '#/components/parameters/StoreScope'
        - name: type
          in: query
          schema:
            type: string
          description: Optional versioned command type filter.
        - name: entity_id
          in: query
          schema:
            type: string
            format: uuid
          description: Optional entity filter.
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CANCELLED]
          description: Optional command status filter.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
          description: >-
            Opaque keyset cursor from a previous response's next_cursor.
            Empty starts newest. Cursors are bound to the exact
            Store+filter scope; concurrent inserts never shift
            already-returned pages (no duplicates, no skips).
      responses:
        '200':
          description: Paginated command history with per-device targets and aggregate state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAdminCommandList'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      summary: Create a typed Store-scoped catalog command (durable intent only)
      operationId: catalogAdminCreateCommand
      security:
        - dashboardSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogAdminCreateRequest'
      responses:
        '200':
          description: "Command persisted with per-device targets snapshotted. Creation is not success: Retail must still apply it."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAdminCommand'
        '400':
          description: Unknown type, cross-Store entity, unknown entity, or invalid payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/catalog-admin/commands/{id}:
    get:
      summary: One catalog command with targets and aggregate state
      operationId: catalogAdminGetCommand
      security:
        - dashboardSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/StoreScope'
      responses:
        '200':
          description: Command detail. Append ?payload=1 to include the immutable payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAdminCommand'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown command or wrong Store scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/dashboard/catalog-admin/commands/{id}/cancel:
    post:
      summary: Cancel a pending catalog command (never rolls back applied work)
      operationId: catalogAdminCancelCommand
      security:
        - dashboardSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/StoreScope'
      responses:
        '200':
          description: Command with updated state. Delivered or applied commands cannot be cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAdminCommand'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Unknown command or cross-Store access (returned as not found).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Command was delivered or applied and cannot be safely cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/dashboard/catalog-admin/categories:
    get:
      summary: Store-scoped Category tree rows
      operationId: catalogAdminCategories
      security:
        - dashboardSession: []
      parameters:
        - $ref: '#/components/parameters/StoreScope'
      responses:
        '200':
          description: Categories with parents, status, online policy, revisions and pending flags.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogAdminCategoryList'
        '401':
          description: Missing or invalid operator session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: FORBIDDEN (missing permission), STORE_FORBIDDEN / STORE_SCOPE_REQUIRED (Store membership), MFA_REQUIRED / MFA_SETUP_REQUIRED (session not FULL) or CSRF_REJECTED (mutations).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: '<device-id>.<credential-id>.<secret-hex> (raw secret shown once at provisioning/rotation)'
      description: >-
        Authorization: Bearer <device_uuid>.<credential_uuid>.<64-hex secret>.
        Never in query parameters. HTTPS in production.
    reportAuth:
      type: http
      scheme: bearer
      bearerFormat: 'REPORTING_API_TOKEN (temporary report-read secret until dashboard user auth)'
      description: >-
        Authorization: Bearer <reporting-api-token>. Report scope only;
        device sync credentials are never valid here. Fail-closed: required
        16+ chars in staging/production; open reports need ALL of
        ENVIRONMENT=development, ALLOW_UNAUTHENTICATED_REPORTING=true, and no
        token. Anything else (including missing ENVIRONMENT) fails startup.
    dashboardSession:
      type: apiKey
      in: cookie
      name: __Host-mlc_session
      description: >-
        Opaque server-side human session (ADR-0053): 256-bit random token,
        stored only as a digest. Cookie __Host-mlc_session (HttpOnly, Secure,
        SameSite=Strict, Path=/) outside development; mlc_session over plain
        HTTP in development. Every dashboard route also requires the FULL
        (MFA-complete) stage, a named permission and server-side Store
        membership; mutations additionally require a same-origin Origin/Referer
        and the X-CSRF-Token header. Device credentials and
        REPORTING_API_TOKEN are never valid here.
...